fix: manage of file stream
